refactor(router): router code refactor
* rename `atc_compat.lua` to `compat.lua`
* add a new file `expressions.lua`
* move lots of logic into `atc.lua`
* rename `get_atc` to `get_expression`
* rename `route_priority` to `get_priority`
* rename `atc_escape_str `to `escape_str`
* remove function `paths_resort`
* transplant #9327 
* transplant #9329 
* transplant #9343
* transplant #9346
* transplant #9357
* transplant #9389
* transplant #9394
* transplant #9451
* transplant #9480
