+++ kong-3.1.0-0.rockspec
@@ -54,7 +54,8 @@ build = {
     ["kong.global"] = "kong/global.lua",
     ["kong.router"] = "kong/router/init.lua",
     ["kong.router.traditional"] = "kong/router/traditional.lua",
-    ["kong.router.atc_compat"] = "kong/router/atc_compat.lua",
+    ["kong.router.compat"] = "kong/router/compat.lua",
+    ["kong.router.expressions"] = "kong/router/expressions.lua",
     ["kong.router.atc"] = "kong/router/atc.lua",
     ["kong.router.utils"] = "kong/router/utils.lua",
     ["kong.reports"] = "kong/reports.lua",
