+++ spec/01-unit/08-router_spec.lua
@@ -1,5 +1,5 @@
 local Router
-local atc_compat = require "kong.router.atc_compat"
+local atc_compat = require "kong.router.compat"
 local path_handling_tests = require "spec.fixtures.router_path_handling_tests"
 local uuid = require("kong.tools.utils").uuid
 
@@ -20,8 +20,8 @@ local function new_router(cases, old_router)
     for _, v in ipairs(cases) do
       local r = v.route
 
-      r.expression = r.expression or atc_compat._get_atc(r)
-      r.priority = r.priority or atc_compat._route_priority(r)
+      r.expression = r.expression or atc_compat._get_expression(r)
+      r.priority = r.priority or atc_compat._get_priority(r)
     end
   end
 
@@ -2087,7 +2087,7 @@ for _, flavor in ipairs({ "traditional", "traditional_compatible", "expressions"
 
       describe("check empty route fields", function()
         local use_case
-        local _get_atc = atc_compat._get_atc
+        local _get_expression = atc_compat._get_expression
 
         before_each(function()
           use_case = {
@@ -2105,35 +2105,35 @@ for _, flavor in ipairs({ "traditional", "traditional_compatible", "expressions"
         it("empty methods", function()
           use_case[1].route.methods = {}
 
-          assert.equal(_get_atc(use_case[1].route), [[(http.path ^= "/foo")]])
+          assert.equal(_get_expression(use_case[1].route), [[(http.path ^= "/foo")]])
           assert(new_router(use_case))
         end)
 
         it("empty hosts", function()
           use_case[1].route.hosts = {}
 
-          assert.equal(_get_atc(use_case[1].route), [[(http.method == "GET") && (http.path ^= "/foo")]])
+          assert.equal(_get_expression(use_case[1].route), [[(http.method == "GET") && (http.path ^= "/foo")]])
           assert(new_router(use_case))
         end)
 
         it("empty headers", function()
           use_case[1].route.headers = {}
 
-          assert.equal(_get_atc(use_case[1].route), [[(http.method == "GET") && (http.path ^= "/foo")]])
+          assert.equal(_get_expression(use_case[1].route), [[(http.method == "GET") && (http.path ^= "/foo")]])
           assert(new_router(use_case))
         end)
 
         it("empty paths", function()
           use_case[1].route.paths = {}
 
-          assert.equal(_get_atc(use_case[1].route), [[(http.method == "GET")]])
+          assert.equal(_get_expression(use_case[1].route), [[(http.method == "GET")]])
           assert(new_router(use_case))
         end)
 
         it("empty snis", function()
           use_case[1].route.snis = {}
 
-          assert.equal(_get_atc(use_case[1].route), [[(http.method == "GET") && (http.path ^= "/foo")]])
+          assert.equal(_get_expression(use_case[1].route), [[(http.method == "GET") && (http.path ^= "/foo")]])
           assert(new_router(use_case))
         end)
       end)
